requirements.txt
flask==2.3.3
flask-cors==4.0.0
pymongo==4.5.0
python-dotenv==1.0.0
google-cloud-vision==3.4.1
google-generativeai==0.3.0
pillow==10.0.1
pytesseract==0.3.10
pdf2image==1.16.3
werkzeug==2.3.7
bcrypt==4.0.1
pyjwt==2.8.0
requests==2.31.0
numpy==1.24.3